Refactored: moved packages to api package
