--- src/main/java/com/example/teamcity/annotations/Optional.java
@@ -1,15 +1,0 @@
-package com.example.teamcity.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-/**
- * The fields with this annotation won't be generated by random or parameterizable value
- * It is needed to specify value manually
- */
-public @interface Optional {
-}
