--- src/main/java/com/example/teamcity/annotations/Random.java
@@ -1,14 +1,0 @@
-package com.example.teamcity.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-/**
- * Fields with this annotation will be filled by random data
- */
-public @interface Random {
-}
