+++ src/main/java/com/example/teamcity/api/annotations/Optional.java
@@ -1,0 +1,15 @@
+package com.example.teamcity.api.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+/**
+ * The fields with this annotation won't be generated by random or parameterizable value
+ * It is needed to specify value manually
+ */
+public @interface Optional {
+}
