+++ src/main/java/com/example/teamcity/api/annotations/Parameterizable.java
@@ -1,0 +1,15 @@
+package com.example.teamcity.api.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+/**
+ * Fields with this annotation will be filled by passed value if the value is passed
+ * otherwise won't be filled
+ */
+public @interface Parameterizable {
+}
