+++ src/main/java/com/example/teamcity/api/models/BaseModel.java
@@ -1,0 +1,5 @@
+package com.example.teamcity.api.models;
+
+public abstract class BaseModel {
+
+}
