+++ src/main/java/com/example/teamcity/api/models/BuildType.java
@@ -1,0 +1,25 @@
+package com.example.teamcity.api.models;
+
+import com.example.teamcity.api.annotations.Parameterizable;
+import com.example.teamcity.api.annotations.Random;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class BuildType extends BaseModel{
+    @Random
+    @Parameterizable
+    private String id;
+    @Random
+    private String name;
+    private Project project;
+    private Steps steps;
+
+}
