+++ src/main/java/com/example/teamcity/api/models/Project.java
@@ -1,0 +1,21 @@
+package com.example.teamcity.api.models;
+
+import com.example.teamcity.api.annotations.Random;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Project extends BaseModel {
+    @Random
+    private String id;
+    @Random
+    private String name;
+    private String locator;
+}
