+++ src/main/java/com/example/teamcity/api/models/Roles.java
@@ -1,0 +1,19 @@
+package com.example.teamcity.api.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Roles extends BaseModel {
+    private List<Role> role;
+
+}
