+++ src/main/java/com/example/teamcity/api/models/Step.java
@@ -1,0 +1,19 @@
+package com.example.teamcity.api.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Step extends BaseModel {
+    private String id;
+    private String name;
+    @Builder.Default
+    private String type = "simpleRunner";
+}
