+++ src/main/java/com/example/teamcity/api/models/TestData.java
@@ -1,0 +1,10 @@
+package com.example.teamcity.api.models;
+
+import lombok.Data;
+
+@Data
+public class TestData {
+    private Project project;
+    private User user;
+    private BuildType buildType;
+}
