+++ src/main/java/com/example/teamcity/api/models/User.java
@@ -1,0 +1,23 @@
+package com.example.teamcity.api.models;
+
+import com.example.teamcity.api.annotations.Random;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class User extends BaseModel {
+    private String id;
+    @Random
+    private String username;
+    @Random
+    private String password;
+    private Roles roles;
+}
+
