+++ src/main/java/com/example/teamcity/api/requests/CrudInterface.java
@@ -1,0 +1,10 @@
+package com.example.teamcity.api.requests;
+
+import com.example.teamcity.api.models.BaseModel;
+
+public interface CrudInterface {
+    Object create(BaseModel model);
+    Object read(String id);
+    Object update(String id, BaseModel model);
+    Object delete(String id);
+}
