--- src/main/java/com/example/teamcity/models/BaseModel.java
@@ -1,5 +1,0 @@
-package com.example.teamcity.models;
-
-public abstract class BaseModel {
-
-}
