--- src/main/java/com/example/teamcity/models/Project.java
@@ -1,21 +1,0 @@
-package com.example.teamcity.models;
-
-import com.example.teamcity.annotations.Random;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Project extends BaseModel {
-    @Random
-    private String id;
-    @Random
-    private String name;
-    private String locator;
-}
