--- src/main/java/com/example/teamcity/models/Step.java
@@ -1,19 +1,0 @@
-package com.example.teamcity.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Step extends BaseModel {
-    private String id;
-    private String name;
-    @Builder.Default
-    private String type = "simpleRunner";
-}
