--- src/main/java/com/example/teamcity/models/TestData.java
@@ -1,10 +1,0 @@
-package com.example.teamcity.models;
-
-import lombok.Data;
-
-@Data
-public class TestData {
-    private Project project;
-    private User user;
-    private BuildType buildType;
-}
