--- src/main/java/com/example/teamcity/models/User.java
@@ -1,23 +1,0 @@
-package com.example.teamcity.models;
-
-import com.example.teamcity.annotations.Random;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class User extends BaseModel {
-    private String id;
-    @Random
-    private String username;
-    @Random
-    private String password;
-    private Roles roles;
-}
-
