--- src/main/java/com/example/teamcity/requests/CheckedRequests.java
@@ -1,22 +1,0 @@
-package com.example.teamcity.requests;
-
-import com.example.teamcity.enums.Endpoint;
-import com.example.teamcity.models.BaseModel;
-import com.example.teamcity.requests.checked.CheckedBase;
-import io.restassured.specification.RequestSpecification;
-
-import java.util.EnumMap;
-
-public class CheckedRequests {
-    private final EnumMap<Endpoint, CheckedBase<?>> requests = new EnumMap<>(Endpoint.class);
-
-    public CheckedRequests(RequestSpecification spec) {
-        for (var endpoint : Endpoint.values()) {
-            requests.put(endpoint, new CheckedBase<>(spec, endpoint));
-        }
-    }
-
-    public <T extends BaseModel> CheckedBase<T> getRequest(Endpoint endpoint) {
-        return (CheckedBase<T>) requests.get(endpoint);
-    }
-}
