--- src/main/java/com/example/teamcity/requests/CrudInterface.java
@@ -1,10 +1,0 @@
-package com.example.teamcity.requests;
-
-import com.example.teamcity.models.BaseModel;
-
-public interface CrudInterface {
-    Object create(BaseModel model);
-    Object read(String id);
-    Object update(String id, BaseModel model);
-    Object delete(String id);
-}
