--- src/main/java/com/example/teamcity/requests/UncheckedRequests.java
@@ -1,21 +1,0 @@
-package com.example.teamcity.requests;
-
-import com.example.teamcity.enums.Endpoint;
-import com.example.teamcity.requests.unchecked.UncheckedBase;
-import io.restassured.specification.RequestSpecification;
-
-import java.util.EnumMap;
-
-public class UncheckedRequests {
-    private final EnumMap<Endpoint, UncheckedBase> requests = new EnumMap<>(Endpoint.class);
-
-    public UncheckedRequests(RequestSpecification spec) {
-        for (var endpoint : Endpoint.values()) {
-            requests.put(endpoint, new UncheckedBase(spec, endpoint));
-        }
-    }
-
-    public UncheckedBase getRequest(Endpoint endpoint) {
-        return requests.get(endpoint);
-    }
-}
