+++ src/test/java/com/example/teamcity/api/BaseTest.java
@@ -1,14 +1,14 @@
 package com.example.teamcity.api;
 
-import com.example.teamcity.generators.TestDataStorage;
-import com.example.teamcity.models.TestData;
-import com.example.teamcity.requests.CheckedRequests;
-import com.example.teamcity.spec.Specifications;
+import com.example.teamcity.api.generators.TestDataStorage;
+import com.example.teamcity.api.models.TestData;
+import com.example.teamcity.api.requests.CheckedRequests;
+import com.example.teamcity.api.spec.Specifications;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.asserts.SoftAssert;
 
-import static com.example.teamcity.generators.TestDataGenerator.generate;
+import static com.example.teamcity.api.generators.TestDataGenerator.generate;
 
 public class BaseTest {
     protected SoftAssert softy;
