+++ src/test/java/com/example/teamcity/api/BuildTypeTest.java
@@ -1,19 +1,17 @@
 package com.example.teamcity.api;
 
-import com.example.teamcity.models.BuildType;
-import com.example.teamcity.models.Project;
-import com.example.teamcity.models.User;
-import com.example.teamcity.requests.CheckedRequests;
-import com.example.teamcity.requests.unchecked.UncheckedBase;
-import com.example.teamcity.spec.Specifications;
+import com.example.teamcity.api.models.BuildType;
+import com.example.teamcity.api.models.Project;
+import com.example.teamcity.api.requests.CheckedRequests;
+import com.example.teamcity.api.requests.unchecked.UncheckedBase;
+import com.example.teamcity.api.spec.Specifications;
 import org.apache.http.HttpStatus;
 import org.hamcrest.Matchers;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
 
-import static com.example.teamcity.enums.Endpoint.*;
-import static com.example.teamcity.generators.TestDataGenerator.generate;
+import static com.example.teamcity.api.generators.TestDataGenerator.generate;
 import static io.qameta.allure.Allure.step;
 
 @Test(groups = {"Regression"})
